Change the contract for append() and prepend()
Previously, bank.append() and bank.prepend() would return the
(potentially very large) array. Mostly because the default
implementation in the Databank class did that.

For small arrays, this was kind of useful, but for bigger ones on
persistent databanks it caused significant hassle.

So, I've changed the contract. If you want to read the whole thing,
read it.